+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Changed
 
+- concord-server: GitHub triggers will now correctly use the repo's
+organization and project names to match the events;
 - concord-server: now process state snapshots can only be accessed by
 admins, process owners and "global readers".